+++ crates/core/src/model/mod.rs
@@ -4,15 +4,16 @@
 #![warn(missing_debug_implementations)]
 #![warn(unused_extern_crates)]
 
-pub mod array;
-pub mod find;
-pub mod object;
-pub mod scalar;
-pub mod value;
+mod array;
+mod find;
+mod object;
+mod scalar;
+mod value;
 
 mod ser;
 
-pub use array::{Array, ArrayView};
-pub use object::{to_object, Object, ObjectView};
-pub use scalar::{Scalar, ScalarCow};
-pub use value::{to_value, State, Value, ValueCow, ValueView};
+pub use array::*;
+pub use find::*;
+pub use object::*;
+pub use scalar::*;
+pub use value::*;